+++ src/main/kotlin/study/step3/race/reader/ActualInputReader.kt
@@ -1,4 +1,4 @@
-package study.step3.race.mock
+package study.step3.race.reader
 
 class ActualInputReader : InputReader {
     override fun raedLine() = readlnOrNull()
