+++ src/main/kotlin/study/step3/race/reader/InputReader.kt
@@ -1,4 +1,4 @@
-package study.step3.race.mock
+package study.step3.race.reader
 
 fun interface InputReader {
     fun raedLine(): String?
