+++ src/test/kotlin/study/step3/race/mock/MockInputReader.kt
@@ -1,5 +1,7 @@
 package study.step3.race.mock
 
+import study.step3.race.reader.InputReader
+
 class MockInputReader(
     private val inputs: List<String>,
 ) : InputReader {
